Use int64 on divide

+++ lang/term_divide.go
@@ -47,9 +47,9 @@ func (p DivideTerm) Eval(ctx *context.Context) (any, error) {
 	}
 
 	switch left := leftEval.(type) {
-	case int:
+	case int64:
 		switch factor := factorEval.(type) {
-		case int:
+		case int64:
 			return float64(left) / float64(factor), nil
 		case float64:
 			return float64(left) / factor, nil
@@ -58,7 +58,7 @@ func (p DivideTerm) Eval(ctx *context.Context) (any, error) {
 		}
 	case float64:
 		switch factor := factorEval.(type) {
-		case int:
+		case int64:
 			return left / float64(factor), nil
 		case float64:
 			return left / factor, nil
